+++ xinference/model/llm/pytorch/core.py
@@ -496,4 +496,6 @@ class PytorchChatModel(PytorchModel, ChatModelMixin):
         else:
             c = self.generate(full_prompt, generate_config)
             assert not isinstance(c, Iterator)
+            if tools:
+                return self._tool_calls_completion(self.model_uid, c, tools)
             return self._to_chat_completion(c)
